Document retained requirements files
