+++ docs/installation/methods.rst
@@ -9,7 +9,7 @@ Manual installation of OWTF is nothing but cloning the repo and running the owtf
     cd owtf/
     make setup
     source /home/{username}/.virtualenvs/owtf/bin/activate
-    python setup.py install
+    python -m pip install .
 
 Docker
 ^^^^^^
@@ -22,3 +22,8 @@ Just make sure that you have ``docker`` and ``docker-compose`` installed and run
 	make compose
 
 * If you wish to override the environment variables for docker setup, use the file named ``owtf.env``
+
+Requirements files
+^^^^^^^^^^^^^^^^^^
+
+The ``requirements`` directory remains the canonical source for pinned dependencies used by builds, documentation, tests, and development environments. The ``pyproject.toml`` file reads these lists dynamically, so the corresponding ``*.txt`` files should stay in the repository even though installation now relies on the PEP 517 workflow (``python -m pip install``).
